Migrate WholeGraphButton to RxJS.
